remove topology_event_handling.go and refactor
